+++ Proyecto9/data.txt
@@ -1,0 +1,18 @@
+8
+2
+-5
+-5
+-2
+0
+0
+4
+3
+5
+0
+-3
+0
+2
+6
+-2
+-7
+6
