+++ Proyecto9/resul.txt
@@ -1,0 +1,1 @@
+[[0, 2, 1, 4, 9, 3, 8, 5, 7, 6], [[0, 2], [[0], ][[2], ]][[1, 4, 9, 3, 8, 5, 7, 6], [[1, 4, 9, 3, 8], [[1, 4, 9], [[1, 4], [[1], ][[4], ]][[9], ]][[3, 8], [[3], ][[8], ]]][[5, 7, 6], [[5, 7], [[5], ][[7], ]][[6], ]]]]
